NEW: Screen captures in doc/
These mostly describe the system at one point in time, often
broken at that time.  For example, 20230828-hdmi-through
shows a fascinating bug in the HDMI--a bug that has since been
fixed.  Likewise, doc/testimg.jpg is the image I've been sending
from the Pi to the test fixture.

The hierarchy and utilization files show how sensitive the
logic usage is to the network table size.